refactor(runtime): rewire demo-bank + examples onto @flowlet/runtime
